mudando a lib do kafka de confluent-kafka para franz-go.
