--- adapter/output/queue/kafka.go
@@ -1,1 +1,0 @@
-package queue
